requirements.txt
acvl-utils==0.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
argparse==1.4.0
arrow==1.3.0
async-lru==2.0.4
asyncio-atexit==1.0.1
attrs==23.2.0
babel==2.15.0
backoff==2.2.1
batchgenerators==0.25
batchgeneratorsv2==0.2
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
deprecated==1.2.14
dicom2nifti==2.4.11
dynamic-network-architectures==0.3.1
einops==0.8.0
fastjsonschema==2.20.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.6.1
future==1.0.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
importlib-metadata==6.11.0
iopath==0.1.10
ipympl==0.9.4
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.11
kiwisolver==1.4.5
lazy-loader==0.4
linecache2==1.0.0
markupsafe==2.1.5
# matplotlib==3.9.1
mistune==3.0.2
monai==1.3.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nibabel==5.2.1
nnunetv2==2.5.1
notebook-shim==0.2.4
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.10.0.84
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.7.0
pandas==2.2.2
pandocfilters==1.5.1
pillow==10.4.0
portalocker==2.10.1
prometheus-client==0.20.0
protobuf==4.25.4
psutil==5.9.8
pycparser==2.22
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
graphviz==0.20.3
python-json-logger==2.0.7
pytz==2024.1
pyyaml==6.0.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
send2trash==1.8.3
simpleitk==2.3.1
sniffio==1.3.1
soupsieve==2.5
sympy==1.13.1
synapseclient==4.4.0
tensorboardx==2.6.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
torch==2.4.0
torchaudio==2.4.0
torchsummary==1.5.1
torchvision==0.19.0
tqdm==4.66.4
traceback2==1.4.0
triton==3.0.0
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
unittest2==1.1.0
uri-template==1.3.0
urllib3==1.26.19
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
wrapt==1.16.0
yacs==0.1.8